When KSPRICHARDSON is used without a KSP monitor and the PC has a PCApplyRichardson() then ksp->its and converged reason are properly set to indicate type of convergence.
  Also with HYPRE BoomerAMG when used as a solver (with KSPRichardson) the KSP convergence
criteria is passed down to BoomAMG and replaces the default (usually 0.0) criteria we normally
pass to BoomerAMG (for when we use BoomerAMG as a preconditioner.

Keita,

    Thanks for pointing this out. This is a "feature" of some of the preconditioners when used with
Richardson. We call the preconditioner with a number of iterations directly and the preconditioner
decides when to exit (rather than all other cases where KSP decides), hence we did not have
information about the number of iterations. So in previous versions we did not set the ksp->its;
petsc-dev was changed to set ksp->its to the maximum number of iterations requested in the
preconditioner.

  This too is not really correct because some preconditioners when used with Richardson
(specifically PCMG and hypre's BoomerAMG) will do the convergence test internally and
may not use the maximum number of iterations requested. I will modify petsc-dev today to
collect this information back from the preconditioner and pass it back correctly to ksp->its.