MatGetSchurComplement
This is the proposed interface for getting preconditioning matrices for
the blocks in PC_FieldSplit.  The rationale is that when performing
block factorization, various Schur complements arise, some of which we
need preconditioners for.  Since the Schur complements are generally
dense, the user should normally provide spectrally equivalent operators
(often as assembled matrices).  Currently the user would implement
MatGetSchurComplement, but I think it would be nice for matrices to keep
a list of user-provided Schur complements to be used for
preconditioning.

The default preconditioning matrix is the SIMPLE operator, i.e. the
matrix inverse which destroys sparsity is replaced by the inverse of the
diagonal.  An alternative would be to use sparse approximate commutators
as in

@article{elman2006bpb,
  title={{Block preconditioners based on approximate commutators}},
  author={Elman, H. and Howle, V.E. and Shadid, J. and Shuttleworth, R. and Tuminaro, R.},
  journal={SIAM Journal on Scientific Computing},
  volume={27},
  number={5},
  pages={1651--1668},
  year={2006},
  publisher={Philadelphia, PA: SIAM, c1993-}
}

This patch adds MAT_IGNORE_MATRIX to MatReuse.  The rationale is that it
may be much more efficient to obtain the real Schur complement (e.g. as
Mat_SchurComplement) and the approximation for preconditioning at the
same time.  An alternative is to make sure the Mat caches enough
information so that this is not a hit, and then add a flag to indicate
whether the real or preconditioning variant is being requested.

--HG--
extra : rebase_source : 58ac5e56058d162782ec8d315f828cf1265f9169