Add support for -da_refine_hierarchy_{x,y,z}
These options accept a list of refinement ratios, starting from the
finest grid.  This is useful for partial semi-coarsening, e.g.

  -da_grid_x 3 -da_grid_y 11 -da_grid_z 11 -nlevels 4 \
    -da_refine_hierarchy_x 1,1,10 \
    -da_refine_hierachy_y 2,2,1 -da_refine_hierarchy_z 2,2,1

produces levels 3x21x21, 3x41x41, 21x41x41.  This sort of scheme would
be expected to work well for a problem with much stronger coupling in
the X direction than in Y or Z, provided the fine-level smoother was
sufficiently strong in the X direction.