Add Tim Davis's ordering package AMD, -pc_factor_mat_ordering_type amd
So we currently install AMD with UMFPACK, since it is a dependency.  In
principle AMD can be used alone, but I'm too lazy to make the build
system recognize AMD as an independent entity (note that many users will
get it from SuiteSparse which includes both of these plus CHOLMOD, KLU,
and other ordering pieces).  So there are ugly references to UMFPACK in
this file, despite the fact that the dependency is only on AMD, not
UMFPACK.

--HG--
extra : rebase_source : bc71cdab8f592cb469e148ac5f85446d794300a5