C-style comments
Preprocessors also remove comments, and some like to complain about //
comments in C89 source, even if they are within #ifdef __cplusplus and
won't be parsed anyway.  Use of C-style comments in headers that will be
included from C (even if none of the header is actually active) avoids
the spurious warning.