Make PetscValidLogicalCollective raise the error on all processes
The old version only raised an error if (mine != global_max), but this
is not useful, especially if rank 0 does not detect an error (since rank
0 is usually responsible for actually printing the error).  This new
version computes a global maximum of [-mine,mine], which computes both
global min and max in one MPI_Allreduce.  The value is logically
collective if (global_min == global_max), so the error can be raised
collectively.