Make -ksp_diagonal_scale use the diagonal from the preconditioning matrix
so that it works when the preconditioning matrix is different from the system matrix.
Perhaps it would be better to take the diagonal from the true matrix if it is provided?
That is tricky if the call will actually fail because it is indirectly not implemented (e.g. with MATNEST).