Namespace a bunch of global variables used for logging
Perhaps these should be placed in a smaller number of thread-local structures.

Note that I used the "petsc_" prefix for these because it's what was
used by nearby variables. These are not public symbols, so it's
inconsequential (in the sense that it can be changed without impacting
source-level compatibility and does not appear in documentation), but
the symbols have extern linkage, so they still need to be namespaced.