Fix MatXAIJSetPreallocation, was conflating block size and over-using array.
Also move call to MatSetBlockSize() because it is safe to call early in
petsc-3.3 and ensures that the local size of the layout is
compatible. Call to PetscLayoutSetUp() will currently be done twice, but
is required before MatGetLocalSize() to ensure that the local size is
divisible by the block size.