Fix nasty bug in PetscTable reallocation
The old code decided whether to reallocate up-front, before checking
whether the new key was already in the hash, but then assumed that it
was not when checking validity of the reallocated hash. The new code
defers the check of hash size until after it has been determined that
the key is not already in the hash. In the (common) case when the key is
already in the hash, the new code is one less conditional, thus likely
also marginally faster.

--HG--
extra : amend_source : f8a6ec21f6a9bac5684a1daf0806f85695cb6ea0